+++ packages/web-workers/src/orchestration/orchestration-telemetry-properties.ts
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+export interface OrchestrationTelemetryProperties {
+    requestResponse?: string;
+    instanceId?: string;
+    isReplaying?: string;
+    currentUtcDateTime?: string;
+    totalWaitTimeInSeconds?: string;
+    activityName?: string;
+    failureMessage?: string;
+    waitEndTime?: string;
+    waitStartTime?: string;
+    scanId?: string;
+}
